More int to double conversions for tempo

+++ include/vrv/doc.h
@@ -220,7 +220,7 @@ public:
     bool ExportTimemap(std::string &output);
     void PrepareJsonTimemap(std::string &output, std::map<double, double> &realTimeToScoreTime,
         std::map<double, std::vector<std::string>> &realTimeToOnElements,
-        std::map<double, std::vector<std::string>> &realTimeToOffElements, std::map<double, int> &realTimeToTempo);
+        std::map<double, std::vector<std::string>> &realTimeToOffElements, std::map<double, double> &realTimeToTempo);
 
     /**
      * Set the initial scoreDef of each page.
